[RPP] Deduplication of track appenders: use old |appendAsyncEventsAtLevel| on sync events
First the appender system is only responsible for rendering the UI, so it does not need to collect information about the call stack of events.

The logic of `appendSyncEventsAtLevel` is to iterate over each level, starting from the beginning. For each level, check if the start timestamp of the event is already used by that level. If yes, continue to the next level. If no, append the event to the level.

This also works for sync events if they are sorted by start time and if the children event's time is fully contained within its parent's time. When we reach a children event, its start time will be equal to or greater than the parent's, so it will be greater than the `lastUsedTimeByLevel`. Therefore, we will append it to the next level.

And since
    - the events should already be sorted by the handlers.
    - the code that generate parent and child event can guarantee parent is always covers the child (chromium.googlesource.com/chromium/src/base/trace_event/common/+/refs/heads/main/trace_event_common.h).
So we can just use the appendAsyncEventsAtLevel on sync events as well

Bug: 1427378
Change-Id: I28fbf687b7eefedf0c07692c6274d1627ed1b007