Migrate long task candy striping to data provider
This CL migrates the candy striping of main thread long tasks out of the
FlameChart component and into the data provider, using the new
decorations system.

This also means the FlameChart no longer dives into the groups it gets
given to render, and keeps the abstraction that it is a chart that
renders events, and knows nothing more.

This also fixes a bug where we were incorrectly candy striping idle
tasks.

Bug: 1434297
Change-Id: If21d0161f093f045329b772990d59f4864ca16cb
Fixed: 1445878