[Accessibility] Add support for showing context menu by keyboard
The root cause if that the onBodyContextMenu and onHeaderContextMenu
limit the event.button to be 2, which works well when user using mouse
but limit the access for the keyboard user. The fix will allow showing
context menu when user using keyboard to imitate right click action,
for example 'Shift + F10' in windows.

Bug: 1446482
Change-Id: Id3ed431f9b3b28517f360dd7c70de1754934b63c