[cleanup] Simplify binding creation in Automapping.
As a first step, turn createBinding() into an async function to make it
easier to follow the control flow and await metadata.

As a second step, remove the redundant sourceCodeToMetadataMap, which
previously was populated first and then read during onMetadatas(). This
map is redundant, because both the ContentProvider based project and the
file system project anyways cache the metadata per UISourceCode already.

Bug: none
Change-Id: Id9726dd973e2bd8deb241fe55305609b6f2290b1