[Workspaces] Fix automatic mapping for files embedded in source maps.
When source files are embedded within source maps (via the
`sourcesContent` field), it's impossible to tell the size of their
content (on disk) in bytes. But still the CompilerScriptMapping was
using the string length of the embedded sources as if it was equal
to the content size in bytes; in most cases this was actually the
case, as long as no non-ASCII characters were used.

Screenshot: https://imgur.com/1rdgXdN.png
Fixed: chromium:1446383
Change-Id: I043ea7f54bc729a4f6c4c1d7943fd4ab327a616d