Remove gzip method from InspectorFrontendHost
The lack of definition for it in devtools_compatability was causing an
error:

```
"Incompatible embedder: method Host.Inspect
orFrontendHost.isGzip is missing. Using stub instead.", source: devtools://devtools/bundled/c
ore/host/InspectorFrontendHost.js (336)
```

Given that we do not want this method to be on the top level, I have
just moved it to within the method that will use it. This has minimal
impact but avoids the error and avoids us putting methods onto
InspectorFrontendHost that we actually don't want to expose (even though
we can use the TS `private` keyword, it's still public really).

Bug: none
Change-Id: I125fdca3db3b6bbc497b6aeb5b4444e809075b4a