[web tests] Remove all web tests from DevTools repository
After migrating the tests to the Chromium repository and removing the
step that runs these tests on the DevTools bot, we can now remove
the web tests from the DevTools frontend.

The initial experiment was to see if it's feasible to host the
DevTools related web tests directly in the DevTools repository. But
we'd lose too much coverage (dfferent platfroms/configs/etc). In the
end we decided to keep DevTools web tests in the Chromium repo and
focused more on DevTools e2e tests.

R=dsv@chromium.org

Fixed: 1447470
Change-Id: I3b6830a57d64fbd3b7351b1d1bdc202165553fea