Intercept `cancel` event on Dialog component
The cancel event is dispatched when the Escape key is hit while the
<dialog> is open. When this happens the original Key event doesn't make
it to the Dialog component, meaning that the handling for Esc strokes in
the Dialog component don't execute when the Esc is hit. This is
problematic because the dialog is closed but the state isn't updated
internally.

This wasn't the case before: when the Esc key was hit, the keydown
event listener was dispatched. I'm unsure at which point this changed,
but the <dialog> element is actively being developed [1] and as such
it's not surprising.

In any case, this CL listens to the `cancel` event in the Dialog
component to make sure the internal states are handled properly when the
Dialog is closed using the Esc key.

[1] https://chromium-review.googlesource.com/q/owner:jarhar@chromium.org+dialog

Fixed: 1447475
Change-Id: Ide8aae9c3634bc0bbc2043e156961a5095ca4a94