Show infobar when ignore listing for any reason
Previously we only showed an infobar when ignore listing based
on the url matching a regex rule. Now we show the infobar when
a file is ignore listed for any reason.

Bug: 1440929


Fix bug, update test


Expanded stack traces with ignore listed frames can be collapsed

In the Console and Applications pane, whenever we provide a link
to show more frames because some are ignore listed, we will show
a "Show less" link after the user expands so they can click it to
collapse the ignore listed frames again.

Fixed: 1364501
Change-Id: I758c96b6da26c71cece917b7cc165887ed67a7ab