[debugger] Add 'ScriptSourceWasEdited' event on DebuggerModel
V8 removes breakpoints after a script was live edited. This requires
us to re-set them from the frontend. This is currently done by the
`ResourceScriptFile` but we want to clean up the layering a bit and
let `Breakpoints` handle this responsibility.

As a first step we need a new event that gets fired after the live
edit is done. Even though the Script fires the event, we emit it
via the DebuggerModel. Script is not an event emitter itself and
subscribing to individual `Script` instances is cumbersome.

R=jarin@chromium.org

Bug: 1444349
Change-Id: If107ec84de98ea19ff169cccd22d5337473df17d