Clarify rendering magic numbers in FlameChart.ts
This CL attempts to deduplicate and tidy up the rendering logic inside
FlameChart.ts which rendered some events with odd pixel values
(subtracting 0.4 px from the width, for example). For some of these
values I have realised why and left a comment (e.g we subtract 1 from
the height so that there is a gap between rows). For others I have
removed the small adjustment and verified locally with a few traces that
it does not make a difference visually.

Bug: 1440169
Change-Id: Ia4f4ffba211e4dce2e33c67026ebc18f2230d65d