Update to use the existing indexed db tree element
Add IndexedDB to Storage Buckets Tree

Adds an IndexedDB tree under each bucket in the storage buckets tree.
Each one displays the IndexedDB information stored in the bucket. The
current top level IndexedDB tree remains, and it continues to display
only the IndexedDB information of the default buckets for all origins.

Bug: 1406017
Change-Id: I1d0404bf849d548e6ca5a9b87baae7ac3ed262e7


Add Storage Buckets to Application panel

Adds the Storage Buckets menu to the sidebar of the Application panel.
The Storage Buckets menu is a tree which contains a list of all the
storage buckets for the page. Under each storage bucket will be  a list
of supported storage APIs. To use, this must be enabled in the
Devtools' experiments settings menu.

Bug: 1406017
Change-Id: I66840c0f12624486c61e15775ebbfb13d8f38c7d