Remove TempBackingStorage
This storage mechanism was used for storing Perf Panel traces but we do
not need it any more and can move away from it.

Longer term there is more we can do here to tidy up the way loading a
file works in the Perf Panel, but as a start this CL removes the
TempBackingStorage that maintained the JSON of the trace. Now that the
file saving is done by encoding the JSON from the new engine
(crrev.com/c/4381791), we do not rely on the backing storage to generate
the JSON that we then write to disk.

Bug: 1428842
Change-Id: Ic9cfd12b8aa7425cfab4b4392520ea03571069e2