Save context before drawing decorations
Also update the whole panel screenshot test to load a more interesting
trace. This test was recently changed to load a basic trace that
didn't render any event in the timeline, with the reasoning that we
wanted to test only loading and have a dedicated test for each track to
test event rendering. However, we have multiple tracks w/o screenshot
test and as such having a full timeline screenshot test will give us
some coverage while we migrate tracks to the new system. This test
should cover us in case a similar issue to the one fixed here is
introduced.

Fixed: 1450190
Change-Id: I6f6f07fd06d7dc3b088de69739b43b43e804bd20