[sources] Don't expand Threads list by default.
Eventually we want to make it possible for developers to reorder the
Sources panel sidebars according to their needs. But in the mean time
we should at least make sure that we optimize for the common use cases,
and the fact that the Threads sidebar pane is expanded by default is
likely not optimized for the common use case.

Fixed: chromium:1318590
Change-Id: Icb54c440879cdeb61c143de15d63ae7d57f425aa