Extract BreakpointManager from 'bindings/' into 'breakpoints/'
This is necessary to be able to call the NameResolver of the
SourceMapScopes module from the BreakpointManager. Otherwise we'd
have a circular dependency.

The CL is a pure moving operation and does not contain any functional
changes.

For web tests we still put the globals onto 'self.Bindings', but
these will go away once we complete crbug.com/1442410.

R=jarin@chromium.org

Bug: 1444349
Change-Id: Id56f27c67cd8c981a16097052e1e2f0037a695b4