Do not count async events in Pie Chart
Recently I noticed that the pie chart in some traces had an entry for
"experience" events (see updated screenshots). I figured this is an
incorrect behaviour because the "experience" events in question are
async and as such their duration cannot be added to the duration of
sync events that ran in a thread to calculate total timings.

This bug was introduced accidentally in https://crrev.com/c/4481135,
where I accidentally passed the flag that prevented async events to be
filtered out in the pie chart aggregation.

Bug: none
Change-Id: I6b8067478684b4fcf58b1e77b0372ea5a9d07470