PreloadingDevTools: Split out "This Page" panel
Before this CL, information about preloading used for the current page
was shown at the bottom of the Preloading panel. We are planning to
call it "This Page" panel and to add more information. As a first step,
this CL split out it into dedicated panel. Adding information is a
future task.

Bug: 1410709
Change-Id: Ia1a80d40a276569ec706a103e001c16f2b8665f8