Avoid relative paths in ESLintRc
If ESLint is run directly on a file nested in a sub directory, it may
not find the right TSConfig file to use (I see this happen in my editor
a lot where the editor's ESLint plugin cannot locate the file as the
path is wrong). Swapping the path to be an absolute path fixes this.

Bug: none
Change-Id: I35aa687df2bebddfd28ced91c4cd96ece0cce7c1