Removes network panel focus when receiving websocket message.
Each time a websocket message comes in, we update the network viewport to re-render with a new timestamp inside the update call. The default is to focus on the selected cell regardless of whether the user is actively somewhere else like in the console.
This patch only limits focus to when the user is actively scrolling the network log view.

Screenshot:
https://i.imgur.com/Ush98Yv.gif

Bug:1449881
Change-Id: I9a71b2ba3e3066a3a8e7e2edc89d4e0b47611bf7