Replace BreakpointsSidebarPane with a legacy wrapper.
This is a pure refactoring: we used to have a BreakpointSidebarPane UI.Widget that didn't do anything other than embedding BreakpointsView Lit component. When user performed some action, BreakpointsView Lit component would send an event, that BreakpointSidebarPane UI.Widget listened to and delegated handling to the BreakpointsSidebarController.

The reason for using events here, I suppose, is that BreakpointsView is in the front_end/panels/sources/components, while BreakpointSidebarPane and BreakpointsSidebarController are in front_end/panels/sources and since sources depend on components, components cannot depend on sources and thus cannot invoke methods directly.

None of this makes much sense to me.

I've implemented a LegacyWrapper UI.Widget previously that can wrap Lit component in a generic way.

This CL does the following:
1) Moves BreakpointsSidebarController with no changes from front_end/panels/sources/BreakpointsSidebarPane.ts to the front_end/panels/sources/components/BreakpointsView.ts file
2) Removes BreakpointsSidebarPane.
2a) Replaces its usage with LegacyWrapper<BreakpointsView>
2b) Makes BreakpointsView call BreakpointsSidebarController directly instead of sending events that used to be handled by BreakpointsSidebarPane
3) Updates the tests of BreakpointsView to account for 2b)
4) Moves the tests of BreakpointsSidebarController from BreakpointsSidebarPane_test.ts to BreakpointsView_test.ts

The diff is so big because 1) and 4).
Here's are diff of BreakpointsSidebarPane.ts and BreakpointsView.ts:
https://diff.googleplex.com/#key=UpVxPmEExc8O
and respective tests: https://diff.googleplex.com/#key=cQh3xsj9zfde



Bug: 1452089
Change-Id: I0b02e58bf93a79e42bacdc5258b3d8920062a156