Enable creation of FilmStripModel.Frames from TraceEngine events
This isn't used yet, but is part of a larger piece of work to migrate
FilmStripModel to the new engine.

Additionally this CL removes the code for looking for CaptureEvent.
Having spoken with mmocny@, paulirish@ and yoavweiss@ to check, we
agreed that this event does not exit any more and has not for a while,
so we can safely drop the code that dealt with this event.

Bug: 1453234
Change-Id: Ia4cca0066a08b40b1b5bde995f8e27d2b09c80f2