Move FilmStrip out of PerformanceModel
This CL moves the instantiation of FilmStrip out of the
PerformanceModel. The motivation here is that we are working on making
the FilmStripModel use the new trace engine's data, and thus do not want
it to live nested within the PerformanceModel.

Instead, we move the instance of FilmStripModel for the Performance
Panel out and define it in TimelinePanel. The rest of this CL is
primarily wiring changes to pass the instance into all the setModel
calls that expect it.

This seems like a lot of effort for not much gain, but moving the
instance of the model up to this level will make it easier to begin
migrating to the new trace engine. I hope in time we can actually remove
the instance of it, and instead just have a helper function that can
take the new trace data and calculate the frames from it (likely with
some caching to avoid mega-recalcs). I will continue on this direction
in follow-up CLs.

Bug: 1453234
Change-Id: I320a2cec4044db8819e3fc52d2dbe180f6284c20