Restructure Handler types
This CL started as an attempt to move the definition of
PartialTraceParseDataDuringMigration...

It should really be defined in Handlers...so this CL does that but also
exposes it on the Model as I think that is cleaner from the outside.

The motivation here is that I want to add models/trace/extras as part of
crrev.com/c/4610230, which will need the ability to take the
PartialTraceParseDataDuringMigration type as a parameter. To import it,
I therefore need it to live in Handlers, otherwise it lives in ModelImpl
and the moment I do:

```
import {PartialTraceParseDataDuringMigration} from '../ModelImpl.js'
```

And add a DEP to BUILD.gn, Ninja fails because of circular dependencies.

But then I hit a bug in the release build - exactly what I am not sure,
but likely something to do with the way that the ModelHandlers.ts file
acted like a standalone module, except it wasn't, and this caused some
handlers to be undefined in Release builds. To work around this and tidy
it up, I therefore removed the additional layer and exposed all handlers
directly in handlers.ts, whilst creating a type that represets the
ModelHandlers that we can use in situations that require it.

Bug: 1453234
Change-Id: I171cf0b1ac5db2b8fba09743566e845938493097