Move PartialTraceParseDataDuringMigration into Handlers
It should really be defined in Handlers...so this CL does that but also
exposes it on the Model as I think that is cleaner from the outside.

The motivation here is that I want to add models/trace/extras as part of
crrev.com/c/4610230, which will need the ability to take the
PartialTraceParseDataDuringMigration type as a parameter. To import it,
I therefore need it to live in Handlers, otherwise it lives in ModelImpl
and the moment I do:

```
import {PartialTraceParseDataDuringMigration} from '../ModelImpl.js'
```

And add a DEP to BUILD.gn, Ninja fails because of circular dependencies.

Bug: 1453234
Change-Id: I171cf0b1ac5db2b8fba09743566e845938493097