Improve BUILD.gn/TS Imports checker script
Over the last few months I have been hacking away on a script that can
take a directory and validate that the list of DEPs in BUILD.gn and all
the gathered JavaScript imports are in sync, which in this case means:

1. Any DEP in BUILD.gn is imported in at least one file
2. Any file imported in a TS source file is listed as a DEP in the
   BUILD.gn

This CL lands the first script that achieves this. There are some edge
cases and corners cut, but I have tested it on a few directories within
DevTools and validated the results.

For example, here is the script running and finding a redundant DEP in
the BUILD.gn which is accurate.

```
$ node scripts/deps/sync-build-gn-imports.js --directory=front_end/models/emulation

Unused front_end/models/emulation/BUILD.gn DEPs
=> 'front_end/core/platform'
```

The script does not recurse, so if we did want to use this on all of
DevTools we would have to go folder by folder. This is clearly not the
best approach, but for now I have limited the script to one directory so
it is easier to run one by one and find any errors - along with manual
effort to validate the findings.

There are also tests which can be run: (these do not run on the bots.)

```
npx mocha scripts/deps/tests
```

At a high level the script does the following:

1. Parses the BUILD.gn file (via regexes, because clang-format formats
   consistently, I did not feel the need to do more for this first
   attempt) and pulls out a list of modules, their sources, and their
   dependencies.
2. For each source file in the directory, parses it and uses the
   TypeScript AST to gather a list of imports.
3. Those imports and DEPs are then compared (after first being converted
   to absolute paths)
4. Any missing or redundant DEPs are then reported and the script
   errors.
5. If the script is successful, it logs nothing and exits cleanly.

Bug: 1402989
Change-Id: I5b3105311e702502e2726cc8d58f54f0e6375c4b