Gather ui/component/doc errors
When we run screenshot tests, any console errors are not surfaced and
they can often give good information as to why a screenshot test failed,
particularly on bots.

This CL gathers errors that any component examples generate and logs
them at the end of a test run to aid debugging.

Bug: none
Change-Id: I5f0e458b30d46c2e4b34c355293f2406d20f514f