Render screenshots in OPP from the new trace engine
Removes the use of SDK.FilmStripModel from OPP in favour of data from
the new engine. Eventually we should use the TrackAppender system for
this, but it is best to migrate the frames and screenshots track
together, and to do that we need to do some work to expose frames data
in the right format in the new trace engine. Therefore for now in order
to continue to make process on the ultimate removal of
SDK.FilmStripModel and also SDK.TracingModel.ObjectSnapshot, it makes
most sense to migrate the screenshots without creating a new
TrackAppender.

Screenshot tests for this are added in crrev.com/c/4624590.

Bug: 1453234
Change-Id: Ia6e954d8bf519dcba0c5d6212bc75d5eb50b1e89