Remove FilmStripView time mode
The FilmStripView component supports two modes: frames, and times.

When looking at the usage of the component, I found that only the frame
view was ever used. Only the network panel uses this component, and it
immediately set the strip into frame mode. Therefore, we can remove the
idea of modes entirely and just have the frame mode be the only mode.

Bug: 1453234
Change-Id: Iacf841f180736154bdb556b90da1aa06352bbf7e