[Reland] Focus handling changes for NetworkLogView and ApplciationPanelTreeElement.
DataGrid assumes that none of its elements receive focus and the request
selection / focus (or pseudo-focus) is handled by the DataGrid itself.

Following CLs added logic for making the singular request elements
focusable for improving accessibility which introduced a few regressions that might be distracting for the users.

crrev.com/c/4064235, crrev.com/c/4178572, crrev.com/c/4497069

This patch fixes the regressions, maintains voice accessibility in the network datagrid and maintains the initial premise of the data grid being the only focusable thing by marking the cells as focusable and resetting the focus to the datagrid itself when a cell is focused

Bug:1449881,1449881
Change-Id: I9a71b2ba3e3066a3a8e7e2edc89d4e0b47611bf7