Do not share weak maps cache across unit tests
Some perf panel unit tests were flaky and we discovered the use of the
same weak maps across unit tests was the root cause. It turned out,
tests became progressively slower because the same cache was shared with
all tests, making the querying time longer and longer.

Instead, different traces hold different caches or, in the case of
timeline data, the cache is reset before parsing a trace in an unit
test,

Bug: none
Change-Id: I7805571af510965274375fcc27a588523bc8a675