Move Outermost target selector feature to inspector_main
We realized we were getting console errors when running DevTools in node
debugging mode (node_app). The errors where about SDK models used by the
outermost target selector feature not being available, which is expected
because the ispected target is node.

To resolve this, we moved the feature registration from main to
inspector_main, which isn't part of the node_app.

Co-authored-by: nancyly@chromium.org
Fixed: none
Change-Id: Icf78cbd62bc2b8ed6d9b0b1367832c3a893fe27d