Show compilation cache kind in performance tool
The v8.compile tracing event now includes the kind of code cache data
that was consumed, so that developers attempting to set up service
workers as described at
https://v8.dev/blog/code-caching-for-devs#use-service-worker-caches can
see whether their changes are working as intended. It might be nice to
expose the same data in devtools so those developers needn't use
chrome://tracing to see the necessary details.

Bug: chromium:1447122
Change-Id: I18a17ab286b6bf876c7e24a15b1252230670a4e3