[scopes] Add ScopeTreeCache class
This CL adds cache for scope trees keyed on scripts. We use
SDK.Script for the key, which boils down to model + script ID.

The cache is necessary since we want to switch over to use the full
scope tree of scripts, and not just parse single functions.

A follow-up CL will hook the cache up to NameResolver.ts.

R=jarin@chormium.org

Bug: 1457874
Change-Id: If79ff59f2cd1fdbce8b7a6a048a9e5ef876583b8