[RPP] Let NetworkHandlerHandler handle the cached event even if they don't have timing data
Currently in TraceEngine, we will drop a request if it has no |timing|
data.
But those cached request could have no |timing| data.
And how we do now is to calculate the |timing| related timestamp first
from |timing|, and if they are cached, we overwrite them.

To solve those problem, this CL does:
- Only calculate the |timing| related data when the request is not
cached.
- Separated the calculated data and the original data from trace event.
So the |args.data| can be served as source of truth.
- Add a temporary test to check the network request is handled same as
OPP.

Bug: 1457485
Change-Id: Idf019955aa067cc64df8a456907a1bb75ec28e6c