Support TraceEngine.Extras.FilmStrip in FilmStripview Dialog
This CL updates FilmStripView.Dialog to be able to run from either the
legacy SDK model or the new Trace Engine model.

Once this lands, I can update the usage of this component (the Network
Panel) to use the new engine, and then I can remove the support for the
old trace engine.

Bug: 1453234
Change-Id: I16bfc82586b2303678447f66ed9b5844d8fb9446