Migrate NetworkPanel to new FilmStrip
The NetworkPanel uses FilmStripView to render screenshots of the loading
of the page. This used the SDK.FilmStripModel model in order to do this.
As part of the migration to the new trace engine, we need to remove all
uses of the old engine, and so this CL does just that. It moves the
NetworkPanel to use the new FilmStrip from TraceEngine.Extras.

The NetworkPanel is also the only place where FilmStripView is used, so
this CL also updates FilmStripView to support only the new
TraceEngine.Extras.FilmStrip, and removes any mention of
SDK.FilmStripModel from it.

Bug: 1453234
Change-Id: I2ceb48b7d23a07c75a8286d6574595cafdd774fe