Disabled navigation on resize/move animations when animations and elements view are in the same location
When resizing or moving animations in the animations view, the view immediately navigates to the elements view. This becomes an issue when the elements view and animations view are in the same location (e.g panel or drawer).  The user is no longer able to resize or move the animation since the view disappears.

Before and After:
https://imgur.com/a/i4ns8Ad

Bug: 1459043
Change-Id: Ifba9b4c256f0ef4760a25d0993e19c32b8d66f1f