[build] Extend devtools_release_sourcemaps to ESBuild bundles
DevTools has two ways to build release builds:
  * tsc + rollup + terser
  * and esbuild.

THe GN arg "devtools_skip_typecheck" decides which of the two
toolchains is used, as the esbuild one is much faster and enough
for the average Chromium developer.

Until now, the "devtools_release_sourcemaps" only works with the
rollup toolchain. This CL implements sourcemaps also for esbuild.

This means that "devtools_release_sourcemaps" now works
independently of "devtools_skip_typecheck".

This is convinient for dogfooding our own source map support in
DevTools: Now we can test against both rollup and esbuild with
DevTools-on-DevTools.

R=jacktfranklin@chromium.org

Bug: 1430396
Change-Id: I54849764a26f3fb539535eccdde7a79f9fd00d5f