[build] Extend devtools_release_sourcemaps to ESBuild bundles
This CL implements release source maps for ESBuild bundles. This is
when "devtools_skip_typecheck" or "devtools_fast_bundle" GN args are
set we'll now also emit a source map for the entrypoint.

This is convinient for dogfooding our own source map support in
DevTools since we now can test against both rollup and esbuild with
DevTools-on-DevTools.

R=jacktfranklin@chromium.org

Bug: 1430396
Change-Id: I54849764a26f3fb539535eccdde7a79f9fd00d5f