Remove references to Target in CPUProfileDataModel & ProfileTreeModel
In order to use this library in the perf panel's new trace engine we
need to extract the library out of SDK because SDK depends on the
trace engine and using the library would create a circular dep. Also,
SDK is a bloated library at this point, depending on it would make the
dream of making the perf panel portable harder.

Unfortunately, the CPU parsing model depends on SDK.Target, so we need
to account for that first, which is this CL's purpose.

The only usage of Target in this library was to store a profile's Target
instance and expose it in a getter. Instead, this CL saves the reference
to the target in the objects that own the CPUProfileDataModel or
ProfileTreeModel instance.

Bug: 1457230
Change-Id: Id4a19f41ff1004e40abcacbb5d09753a7093780f