Migrate Frame Preview to new FilmStrip
This CL is the other place where the FilmStripView>Dialog component is
used to show a preview of a frame and allow the user to click it to view
it larger. It relied on the SDK.FilmStripModel and therefore needed to
be migrated to the new film strip to unblock the removal of
SDK.FilmStripModel.

Bug: 1453234
Change-Id: Ia329649adeb22ddcaf554dd14988496947157fba