Extract lines from CPU profile coming from trace
This property enables the feature to provide line level execution time
of a function in the sources panel.

Unlike the CDP.Profiler.Profile type, the profile coming in trace events
contains the data as part of the profile itself, making its type shape
different from that of CDP. For this reason a new type is introduced in
the CPU Profile parsing library, which marks this property as optional.

Bug: 1457230


[RPP] Export a CPUProfile with CDP format for each profile in trace

In preparation to invoke the CPU Profile parsing library from the
handler

Bug: 1457230
Change-Id: I8db7ae63012c465c2e2bf6366ab564488b4c77ca